% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{defaultRateFunctions}
\alias{defaultRateFunctions}
\title{defaultRateFunctions}
\format{\preformatted{List of 6
 $ reproFunc      :function (x, time, species, strain)  
 $ deathFunc      :function (stage, x, time, species, strain)  
 $ durationFunc   :function (stage, x, time, species, strain)  
 $ immigrationFunc:function (stage, x, time, species, strain)  
 $ develFunc      : NULL
 $ emigrationFunc :function (stage, x, time, species, strain)  
}}
\usage{
defaultRateFunctions
}
\description{
These default implementations will simply generate errors when
run. To create implementations, please see the documentation linked
to below.
}
\details{
The list should contain the following names, each mapped to a function of the correct signature.

\itemize{
\item \code{reproFunc}
\item \code{deathFunc}
\item \code{durationFunc}
\item \code{immigrationFunc}
\item \code{develFunc}. Note that by defaul develFunc is NULL as it is not required for all simulation types
}
}
\seealso{
\code{\link{RateFunctions}}
}
\keyword{datasets}

