% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivDede.R
\name{derivDede}
\alias{derivDede}
\title{derivDede}
\usage{
derivDede(t, y, parms)
}
\arguments{
\item{t}{Time step}

\item{y}{state variables}

\item{parms}{Additional parameters passed in from the original invoker of the solver}
}
\value{
the new model state.
}
\description{
Internal-use function which allows use of the DeSolve ODE solver. Called by \code{\link{popModel}}
}
\seealso{
\code{\link{derivPBS}}
}
