% Generated by roxygen2 (4.0.1): do not edit by hand
\name{durationFuncDefault}
\alias{durationFunc}
\alias{durationFuncDefault}
\alias{durationFunction}
\title{Life stage duration function}
\usage{
durationFuncDefault(stage, x, time, species, strain)
}
\arguments{
\item{stage}{(integer). The numbered life stage for which to return a duration}

\item{x}{Vector of state variables within the DDE solver. To access a particular variable use:
x$speciesName['stageName',strainNumber]
e.g. for species 'Bacteria', stage 'reproductive', strain 2 use
x$Bacteria['reproductive',2]
If there is only one stage and strain in species 'Food', for example, use
x$Food[1,1]}

\item{time}{(scalar). The current time point in the DDE solver.}

\item{species}{(integer). The numbered species for which to return a life stage duration.}

\item{strain}{(integer). The numbered strain for which to return a life stage duration.}
}
\value{
Duration of the life stage for the stage, species, strain and time specified.
}
\description{
Return the duration of the life stages (time units).
}

