% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericPlot.R
\name{genericPlot}
\alias{genericPlot}
\title{Generic Plotting Function}
\usage{
genericPlot(
  output,
  numSpecies,
  numStages,
  speciesNames,
  stageNames,
  saveFig = FALSE,
  figType = "eps",
  figName = "stagePopFig"
)
}
\arguments{
\item{output}{Model output, from \code{\link{popModel}}}

\item{numSpecies}{Number of species in the model}

\item{numStages}{Number of life stages}

\item{speciesNames}{species names (vector of strings). Default is NULL.}

\item{stageNames}{stage names (a list of vectors if there is more than one species).}

\item{saveFig}{Choose to save the figure (TRUE or FALSE). Default is FALSE.}

\item{figType}{Figure format can be 'eps', 'tiff' or 'png'. Default is 'eps'}

\item{figName}{filepath to save figure to. Default is 'stagePopFig'}
}
\description{
Plot the state variables over time (sums over all strains if there are multiple strains in a species)
}
\seealso{
\code{\link{popModel}}
}
