% Generated by roxygen2 (4.0.1): do not edit by hand
\name{immigrationFuncDefault}
\alias{immigrationFunc}
\alias{immigrationFuncDefault}
\alias{immigrationFunction}
\title{Immigration Function}
\usage{
immigrationFuncDefault(stage, x, time, species, strain)
}
\arguments{
\item{stage}{(integer). The numbered life stage for which to return an immigration rate.}

\item{x}{Vector of state variables within the DDE solver. To access a variable use:
x$speciesName['stageName',strainNumber]
e.g. for species 'Bacteria', stage 'reproductive', strain 2 use
x$Bacteria['reproductive',2]
If there is only one stage and strain in species 'Food', for example, use
x$Food[1,1]}

\item{time}{(scalar). The current time point in the DDE solver.}

\item{species}{(integer). The numbered species for which to return an immigration rate.}

\item{strain}{(integer). The numbered strain for which to return an immigration rate.}
}
\value{
Immigration rate (amount per unit time) for the strain, stage, species and time specified.
}
\description{
Return the rate of immigration (amount per unit time)
}

