% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateFuncCheck.R
\name{rateFuncCheck}
\alias{rateFuncCheck}
\title{rateFunCheck}
\usage{
rateFuncCheck(
  x,
  numSpecies,
  numStages,
  numStrains,
  timeDependDuration,
  rateFunctions,
  speciesNames,
  stageNames,
  startTime
)
}
\arguments{
\item{x}{Vector of state variables within the DDE solver. To access a variable use: x$speciesName['stageName',strainNumber]}

\item{numSpecies}{Number of species}

\item{numStages}{Number of life stages}

\item{numStrains}{Number of strains for each species}

\item{timeDependDuration}{vector of logicals defining whether the stage duration is time dependent for each species}

\item{rateFunctions}{List of rate functions to check}

\item{speciesNames}{Vector of species names}

\item{stageNames}{Vector of species' stage names}

\item{startTime}{First entry in timeVec vector}
}
\value{
TRUE/FALSE - TRUE if rate functions are OK, FALSE if not.
}
\description{
check the user-defined rate functions exist and are in the correct format
}
\details{
checks existence of deathFunc, reproFunc, durationFunc, immigrationFunc and develFunc if 
there is time dependent duration.

Also checks that output from rate functions is a single value
}
