% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateFunction.R
\name{RateFunctions}
\alias{RateFunctions}
\title{Rate Functions}
\description{
The rate functions are user-defined functions which give the rates for each life
stage.
}
\details{
These are passed into the \code{\link{popModel}} function via the \code{rateFunctions} argument. 

The signatures for each rate function are defined in their respective documentation files
}
\seealso{
\code{\link{deathFunction}} \code{\link{durationFunction}} \code{\link{reproductionFunction}} \code{\link{developmentFunction}} \code{\link{immigrationFunction}} \code{\link{emigrationFunction}}
}
