% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stagePop-package.R
\docType{package}
\name{stagePop-package}
\alias{stagePop-package}
\alias{stagePop}
\title{Stage-Structured Population modelling}
\description{
stagePop can be used to model the dynamics and interactions of stage-structured populations (i.e. where the life cycle consists of distinct stages - e.g. eggs, juveniles and reproductive adults) in continuous time. It can be used to model predator-prey interactions, host-parasitoid interactions, resource competition and the effects of environmental change on stage-structured species. The model is based on the formulation described by RM Nisbet and WSC Gurney in 1983 in the paper 'The systematic formulation of population models for insects with dynamically varying instar duration. Theoretical Population Biology, 23, 114-135'.  It uses delay differential equations which are solved using the R-packages 'deSolve' or 'PBSddesolve'.
}
\references{
To be done
}
\author{
Helen Kettle
}
\keyword{package}
\keyword{programming}
