% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultRateFunctions.R
\docType{data}
\name{defaultRateFunctions}
\alias{defaultRateFunctions}
\title{defaultRateFunctions}
\format{
An object of class \code{list} of length 6.
}
\usage{
defaultRateFunctions
}
\description{
These default implementations will simply generate errors when
run. To create implementations, please see the documentation linked
to below.
}
\details{
The list should contain the following names, each mapped to a function of the correct signature.

\itemize{
\item \code{reproFunc}
\item \code{deathFunc}
\item \code{durationFunc}
\item \code{immigrationFunc}
\item \code{develFunc}. Note that by defaul develFunc is NULL as it is not required for all simulation types
}
}
\seealso{
\code{\link{RateFunctions}}
}
\keyword{datasets}
